Collapse store implementation into value
